# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "藏药小程序API",
    "description": "藏药电商小程序后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/product/list": {
      "get": {
        "tags": ["商品模块"],
        "summary": "商品列表",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {"type": "integer"}
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {"type": "string"}
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/product/detail/{id}": {
      "get": {
        "tags": ["商品模块"],
        "summary": "商品详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/product/recommend": {
      "get": {
        "tags": ["商品模块"],
        "summary": "推荐商品",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/coupon/my": {
      "get": {
        "tags": ["优惠券模块"],
        "summary": "我的优惠券",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/coupon/receive": {
      "post": {
        "tags": ["优惠券模块"],
        "summary": "领取优惠券",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "couponId": {"type": "integer"}
                },
                "required": ["couponId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/coupon/use": {
      "post": {
        "tags": ["优惠券模块"],
        "summary": "使用优惠券",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCouponId": {"type": "integer"}
                },
                "required": ["userCouponId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/payment/create": {
      "post": {
        "tags": ["支付模块"],
        "summary": "创建支付（Mock）",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderId": {"type": "integer"},
                  "paymentMethod": {"type": "string"}
                },
                "required": ["orderId", "paymentMethod"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/payment/status/{paymentId}": {
      "get": {
        "tags": ["支付模块"],
        "summary": "查询支付状态（Mock）",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/payment/notify": {
      "post": {
        "tags": ["支付模块"],
        "summary": "支付回调（Mock）",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/upload/image": {
      "post": {
        "tags": ["文件上传"],
        "summary": "上传单图（Mock）",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/upload/images": {
      "post": {
        "tags": ["文件上传"],
        "summary": "批量上传（Mock）",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "data": {
            "type": "object",
            "properties": {
              "total": {"type": "integer"},
              "pageNum": {"type": "integer"},
              "pageSize": {"type": "integer"},
              "list": {"type": "array"}
            }
          }
        }
      },
...
    }
  }
}